params from locals and execute in batches
params_dict_from_locals allows for shorter function definitions by
constructing a params_dict for a request directly from the locals
(which consist of passed function arguments).

execute_in_batches is an abstraction helpful in cases where VK API
limits the size of inputs for a method. When a user makes a request
with more inputs than the underlying VK API method allows, the requests
are performed in batches and concatenates the results.